backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
chromadb==0.5.5
openai==1.43.0
